Revert renaming scripting plugin jar:
it conflicts with the compilation on daemon, so transistion should
be planned accordingly, and now it is not a good time for it.
The most important part of the renaming remains intact.

Partially reverts commit "Rename scripting libs and plugin - invert embeddable suffix"

Original commit: dec8eb789974d30e32a4e5bfa294e2c5971b38e9

GitOrigin-RevId: cf8bd0150df5622c456d102b1679a8c83792684e