Support Android Studio
KTIJ-125

This commit just migrates some as bunches from the old repo.
Android Studio runs using same plugin IJ uses

Investigation on all found bunches:
```
AllOpenMavenProjectImportHandler.kt.as41                This EP is not fired by AS
AbstractMavenImportHandler.kt.as41                      This EP is not fired by AS
NoArgMavenProjectImportHandler.kt.as41                  This EP is not fired by AS
KotlinSerializationMavenImportHandler.kt.as41           This EP is not fired by AS
SamWithReceiverMavenProjectImportHandler.kt.as41        This EP is not fired by AS
plugin.xml.as41                                         Adopted
NewMultiplatformProjectImportingTest.kt.as41            This is in tests. We don't have AS tests
HierarchicalMultiplatformProjectImportingTest.kt.as41   This is in tests. We don't have AS tests
PackagePrefixImportingTest.kt.as41                      This is in tests. We don't have AS tests
MultiplatformTestCompatibilityMatrix.kt.as41            This is in tests. We don't have AS tests
MultiplatformProjectImportingTest.kt.as41               This is in tests. We don't have AS tests
GradleConfiguratorPlatformSpecificTest.kt.as41          This is in tests. We don't have AS tests
GradleInspectionTestCompatibilityMatrix.kt.as41         This is in tests. We don't have AS tests
GradleFacetImportTest.kt.as41                           This is in tests. We don't have AS tests
git4idea.xml.as41                                       Bunched for AS as legacy of 181 platform: 0103c0d2fd8c9ae9e7f41898064407a6278c5831
android-lint.xml.as41                                   What's bad for UAST in AS?
MavenProjectImporter.kt.as41                            Disabled with if
GooglePluginUpdateVerifier.kt.as41                      Ported
verify.kt.as41                                          Ported
MLCompletionForKotlin.kt.as41                           Code in this bunch is unused
dslStyleIcon.kt.as41                                    Bunched by mistake for AS as legacy of 182 platform
testResourceBundle.kt.as41                              configuring additions EP for fixing test failures. We don't run tests with AS => no need to configure additional EP
IDESettingsFUSCollector.kt.as41                         Disabled with if
ITNReporterCompat.kt.as41                               This bunch doesn't compile
KotlinExplicitMovementProvider.kt.as41                  Same as ./idea/resources/META-INF/git4idea.xml.as41
KotlinFUSLogger.kt.as41                                 Disabled with if
ProjectConfigurationCollector.kt.as41                   Disabled with if
MyKotlinTest.kt.as41                                    testData. We don't have AS tests
MyKotlinTest.kt.as41                                    testData. We don't have AS tests
ideaPlatform.kt.as41                                    This bunch doesn't compile
```

GitOrigin-RevId: b7193060800ed237c0bc251b05b9cee0276dc712