ModuleInfo caches: avoid library duplication
New IJ workspace model changes equality logic for libraries.
Libraries in dependencies of different modules from now on are considered different.
Multiple unequal LibraryInfo are created for the library with the same roots because of this.
This implies high memory consumption caused by the enormously high amound of resolvers for library modules.

To mitigate this issue for the time being the LibraryInfo euality uses Library#hasSameContent compatibility method and
hashCode repeats the old name-and-root based hashing.

^KT-42607 In Progress

GitOrigin-RevId: 4bce47c9b5de2f0a1ff6de9b188bca187bcc6653