Moved import tests required Android Plugin into a new module. KT-43115
We need to run import tests with the Android plugin in the IDEA.
If just add a dependency to it, then other tests begin to fail
due to the fact that the classpath begins to leak. It was decided
to move these tests into a separate module and make a separate
configuration for them.

Rename modules names

GitOrigin-RevId: 0e593f9eb06930596050103e9b59e3d1426c383c