Moved import tests required Android Plugin into a new module. KT-43115
We need to run import tests with the Android plugin in the IDEA.
If just add a dependency to it, then other tests begin to fail
due to the fact that the classpath begins to leak. It was decided
to move these tests into a separate module and make a separate
configuration for them.

Rename modules names

GitOrigin-RevId: 42dee2797cadbe9d33082d08295e4c97d4789af3