Always render modifiers in key generation
Some descriptor modifiers such as 'operator', for instance, are not rendered if any of the overridden declarations already has a such modifier.
We rather prefer a consistent behavior here so we will render the descriptors identically not matter what.

This fixes a failing test: org.jetbrains.kotlin.idea.resolve.ReferenceResolveTestGenerated.DelegatedPropertyAccessors.InStandardLibrary.testNotNull().

GitOrigin-RevId: a95604222962cb9e6566e402741aeafee61b1345