Reduce calls to ParcelizeAvailability.isAvailable
ParcelizeAvailability.isAvailable invokes project system APIs
that can be slow especially in multi-module projects. When
highlighting a medium-size Kotlin file in a large project I see
around 100-500 ms spent in this method.

Instead of querying ParcelizeAvailability.isAvailable for each
call/declaration, we can query it once when configuring the module.