Remove JvmBytecodeBinaryVersion from the compiler code
Move it to build-common where it's still used in incremental compilation
caches, and simplify a bit. In the future, it'll make sense to remove it
completely and to avoid writing it to caches. In this commit, I don't do
that to prevent the IC cache version to be updated, causing rebuilds for
all JPS projects.

 #KT-41758

Original commit: f63ffc51ae6ce0a3ae5b278f86f49e2e5efbfc33

GitOrigin-RevId: 5d85bbc6ec684d5411b1e3092591307b88eb80df