Better separation between Kotlin and Android IDE plugins
The existing support for Android single-variant sync in Kotlin and
KAPT Gradle model builders is based on assumption that when Android IDE
plugin is loaded it requests `KoltinGradleModel` and `KaptGradleModel`
for all Gradle projects in the build. This is not desired since this
way Koltin plugin relies on the correct behavior of Android plugin even
for non-Android projects.

This change addresses this by shifting the responsibility to request
models for non-Android projects back to Kotlin plugin. This results in
some inefficiency (requesting such models twice) which we will be
addressed in Android plugin/Studio.

The final design that this change implements is:

* Kotlin plugin always requests Kotlin/Kapt models for non-Android
  modules (com.android.base Gradle plugin is not applied)

* Koltin plugin requests Kotlin/Kapt models for Android modules only
  when an old Android plugin is loaded or there is no Android plugin
  (new Android plugin versions communicate their presence through user
  data attached to `resolverCtx`).

* Kotlin plugin can build variant specific Kotlin/Kapt models for
  Android modules when requested by the Android IDE plugin.

GitOrigin-RevId: 68b2999db1a3ccf795c9383d1a01fd0d93ab9baf