[kotlin] Fix visible variables in the debugger on Dex VMs
NOTE: This commit has been partially cherry-picked from
aaeb58959d5b6f09b31817fdfdcccdc5c84b25f6 of intellij/master.
There's no need to port it back. -- Yan

For inline functions, the Kotlin JVM debugger needs to reconstruct the
currently visible variables out of metadata which the Kotlin compiler
encodes in both the order and names of local variables.

On a dex VM, the order of local variables is not reliable, since locals
are kept in registers which may have to be spilled, which resets their
start offsets. This commit introduces a workaround to recover the
actual variable introduction order when running on a dex VM and also
makes the code for reconstructing inline call stacks more robust.

GitOrigin-RevId: 87850672aca4dacd14abca6fe6eb9066a30eb51c