[kotlin] Narrowed string replacement in new project wizard test
`AbstractBuildFileGenerationTest` replaces the actual Kotlin version
string with a placeholder to make comparison with the expected data
version-independent. After the release of Kotlin 1.6.0, the replacement
pattern coincided with the version string of an unrelated package in
`pom.xml`, that resulted in unnecessary replacement and comparison
failure.

This commit changes the unconditional string replacement to a
regex-based replacement that takes the context into consideration.

IJ-CR-17392

(cherry picked from commit 3ba6e17f8a07568085efc5dcab902db8bc699780)

GitOrigin-RevId: 1cd8dacd548c88d15976a4bac2066e3a4315f4c2